Model Update with Enemy AI
I have finished the models that needed completing (Minus the campfire). Managed to create a wondering AI for the zombie. It will also attack a player when it is within range. Currently set to the cyclops using the villager tag and a villager script. I also added the animation for the zombie. Idle, Walking and attacking. Some minor bugs to fix like the zombie walking glitch